requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
sqlalchemy==2.0.35
pydantic==2.9.2
python-jose==3.3.0
passlib==1.7.4
requests==2.32.3
pandas==2.2.3
scikit-learn==1.5.2
pydeck==0.9.1